+++ pravega-sensor-collector/src/main/java/io/pravega/sensor/collector/rawfile/EventGenerator.java
@@ -22,6 +22,8 @@ import org.apache.commons.lang3.tuple.Pair;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import io.pravega.sensor.collector.util.PravegaWriterEvent;
+
 /**
  * Generate Event from file  
  */
