+++ pravega-sensor-collector/src/test/java/io/pravega/sensor/collector/parquet/EventGeneratorTests.java
@@ -9,6 +9,7 @@ import org.slf4j.LoggerFactory;
 
 import io.pravega.sensor.collector.util.FileNameWithOffset;
 import io.pravega.sensor.collector.util.FileUtils;
+import io.pravega.sensor.collector.util.PravegaWriterEvent;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -33,16 +34,4 @@ public class EventGeneratorTests {
         Assert.assertEquals(parquetData.length(), (long) nextSequenceNumberAndOffset.getRight());
     }
 
-    // @Test
-    // public void TestEmptyFile() throws IOException {
-    //     final EventGenerator eventGenerator = EventGenerator.create("routingKey1",10);
-    //     final List<FileNameWithOffset> files = FileUtils.getDirectoryListing("../parquet-file-sample-data/","parquet");
-    //     File parquetData= new File(files.get(0).fileName);
-    //     final CountingInputStream inputStream = new CountingInputStream(new FileInputStream(parquetData));
-    //     final List<PravegaWriterEvent> events = new ArrayList<>();
-    //     Pair<Long, Long> nextSequenceNumberAndOffset = eventGenerator.generateEventsFromInputStream(inputStream, 100, events::add);
-    //     log.info("events={}", events);
-    //     Assert.assertEquals(100L, (long) nextSequenceNumberAndOffset.getLeft());
-    //     Assert.assertEquals(parquetData.length(), (long) nextSequenceNumberAndOffset.getRight());
-    // }
 }
